+++ mapbox_tilesets/__init__.py
@@ -1,3 +1,3 @@
 """mapbox_tilesets package"""
 
-__version__ = "1.0.1"
+__version__ = "1.1.0.dev0"